Ecommerce/accounts/accounts/requirements.txt
cryptography==2.1.3
flask==0.10.1


gevent==1.2.2
requests==2.5.1
wsgiref==0.1.2
flask-sqlalchemy==2.1
gunicorn==19.7.1
pyjwt==1.5.3
sqlalchemy-migrate==0.10.0
